requirements.txt
python-dotenv
django
djangorestframework
markdown
django-filter
